requirements.txt
# Pinned set validated together (Python 3.12): LangGraph 1.x + former ai-service stack.
fastapi==0.135.3
uvicorn[standard]==0.44.0
starlette==1.0.0
langgraph==1.1.6
langchain-core==1.2.28
langchain-ollama==1.1.0
langchain-openai==1.1.12
langchain-google-genai==4.2.1
python-dotenv==1.2.2
openai==2.31.0
pydantic==2.12.5
pydantic-settings==2.13.1
requests==2.33.1
httpx==0.28.1
google-genai==1.72.0
google-auth==2.49.2
yfinance==1.2.1
qdrant-client==1.17.1
sentence-transformers==5.4.0
pika==1.3.2
vnstock==3.5.1
pandas==3.0.2
numpy==2.4.4
